+++ src/tps/TWSPack.cpp
@@ -1,3 +1,48 @@
 using namespace geode::prelude;
 
 #include "headers/TWSPack.hpp"
+#include "../ui/headers/TWSPackCell.hpp"
+#include "../header/boobs.hpp"
+
+void TWSPack::downloadTP() {
+    isDownloading = true;
+    boobs::downloading.push_back(this);
+    auto weak = geode::WeakRef(this); // weakref is fucking bullshit god GEODE IS GOING TO MAKE ME KILL MYSELF
+
+    auto req = geode::utils::web::WebRequest();
+
+    req.onProgress([this, weak](geode::utils::web::WebProgress const& progress) {
+        if (downloadingIndicator && weak.valid()) {
+            downloadingIndicator->setValue(progress.downloadProgress().value_or(0.69420f) / 100.0f);
+        }
+
+    });
+
+    m_downloadTPlistener.spawn( 
+        req.get(this->DownloadURL),
+        [this, weak](geode::utils::web::WebResponse res) {
+            if (!weak.lock()) {
+                return;
+            }
+
+            if (res.ok()) {
+                if(res.into(fmt::format("{}/packs/{}.zip", Loader::get()->getInstalledMod("geode.texture-loader")->getConfigDir(), TPName))) {
+                    Notification::create("Download Complete!", CCSprite::createWithSpriteFrameName("GJ_completesIcon_001.png"));
+                    downloadSuccessful = true;
+                    isDownloading = false;
+                    boobs::downloading.erase(std::remove(boobs::downloading.begin(), boobs::downloading.end(), this), boobs::downloading.end()); // texture workshop pack sacrificing itself to kill cheeseworks
+                    std::string versionSaveThing = fmt::format("{}", TPName);
+                    Mod::get()->setSavedValue<std::string>(versionSaveThing, TPVersion);
+                    if (cell) {
+                        cell->updateDownloadStata();                    
+                    } else {
+                        this->release(); 
+                        // no it seemed like a cool idea at the time
+                        // but it was costing me a shitload
+                        // so i had to release them back into the wild
+                    }
+                }
+            }
+        }
+    );
+}